Removing all .drush.inc files only in Drupal 11

+++ aegir/tools/bin/fix-drupal-platform-permissions.sh
@@ -71,9 +71,11 @@ find ${drupal_root}/{modules,themes,libraries,includes,misc,profiles,core} -type
 printf "Setting permissions of all codebase files inside "${drupal_root}"...\n"
 find ${drupal_root}/{modules,themes,libraries,includes,misc,profiles,core} -type f -exec chmod 0664 {} \;
 
-printf "Removing all .drush.inc files inside codebase "${drupal_root}"...\n"
-find ${drupal_root}/modules/contrib -type f -name "*.drush.inc" -exec rm -f {} \;
-find ${drupal_root}/sites/*/modules -type f -name "*.drush.inc" -exec rm -f {} \;
+if [ -e "${drupal_root}/core/modules/workspaces_ui" ]; then
+  printf "Removing all .drush.inc files inside codebase "${drupal_root}"...\n"
+  find ${drupal_root}/modules/contrib -type f -name "*.drush.inc" -exec rm -f {} \;
+  find ${drupal_root}/sites/*/modules -type f -name "*.drush.inc" -exec rm -f {} \;
+fi
 
 if [ -e "${drupal_root}/vendor" ]; then
   printf "Setting permissions of all codebase directories inside "${drupal_root}/vendor"...\n"
